create documents, moved schemas, methods, publications, update imports, ** schemas not attached
