+++ imports/api/@/methods/chat-methods.js
@@ -2,6 +2,7 @@ import { Meteor } from 'meteor/meteor';
 import { getHost } from '../shared';
 import { isContributorOrAdmin } from '../../@users/user.roles';
 import Hosts from '../../@hosts/host';
+import Processes from '../../processes/process';
 
 Meteor.methods({
   createChat(contextName, contextId) {
