+++ imports/api/@/publish.js
@@ -36,46 +36,11 @@ Meteor.publish('activities', function (onlyPublic = false) {
   }
 });
 
-Meteor.publish('processes', function () {
-  const userId = Meteor.userId();
-  const host = getHost(this);
-  // Processes._ensureIndex({ host, isPublished: true });
 
-  const fields = {
-    title: 1,
-    readingMaterial: 1,
-    imageUrl: 1,
-    meetings: 1,
-    adminUsername: 1,
-    adminId: 1,
-  };
-  if (userId) {
-    (fields.members = 1), (fields.peopleInvited = 1);
-  }
-
-  return Processes.find(
-    {
-      host,
-      isPublished: true,
-    },
-    {
-      fields,
-      sort: { creationDate: 1 },
-    }
-  );
-});
 
 Meteor.publish('activity', function (id) {
   const host = getHost(this);
   return Activities.find({
-    host,
-    _id: id,
-  });
-});
-
-Meteor.publish('process', function (id) {
-  const host = getHost(this);
-  return Processes.find({
     host,
     _id: id,
   });
