+++ imports/ui/RouterComponents/processes/ProcessesListContainer.js
@@ -1,5 +1,6 @@
 import { withTracker } from 'meteor/react-meteor-data';
 import ProcessesList from './ProcessesList';
+import Processes from '../../../api/processes/process';
 
 export default ProcessesListContainer = withTracker((props) => {
   const processesSubscription = Meteor.subscribe('processes');
