+++ lib/collections.js
@@ -3,7 +3,7 @@ import { Mongo } from 'meteor/mongo';
 // Hosts = new Mongo.Collection('hosts');
 
 Activities = new Mongo.Collection('activities');
-Processes = new Mongo.Collection('processes');
+// Processes = new Mongo.Collection('processes');
 // Works = new Mongo.Collection('works');
 // Pages = new Mongo.Collection('pages');
